ANDROID: sched/walt: make walt_ktime_suspended __read_mostly
Most walt variables in hot code paths are __read_mostly and grouped
together in the .data section, including these variables that show
up as frequently accessed in gem5 simulation: walt_ravg_window,
walt_disabled, walt_account_wait_time, and
walt_freq_account_wait_time.

The exception is walt_ktime_suspended, which is also accessed in many
of the same hot paths.  It is also almost entirely accessed by reads.
Move it to __read_mostly in hopes of keeping it in the same cache line
as the other hot data.

Signed-off-by: Todd Poynor <toddpoynor@google.com>
Signed-off-by: Mohammad Hasan Keramat J <ikeramat80@gmail.com>